Add lint-staged configuration and update TypeScript and Babel configurations
