+++ package.json
@@ -13,13 +13,12 @@
         "build": "npm run type-check && npm run webpack -- --env mode=production",
         "start": "npm-run-all --parallel type-check:watch dev",
         "dev": "npm run webpack-dev-server -- --env mode=development",
-        "test": "npx jest",
-        "code-coverage": "jest --coverage",
         "prettier": "npx prettier",
         "prettier:format": "npm run prettier -- --write src/**/*.ts",
         "prettier:check": "npm run prettier -- -c ./src/**/*.ts",
-        "lint-check": "npx eslint ./src/**/*.ts",
-        "code-style-check": "npm run lint-check && npm run prettier:check"
+        "lint-check": "npx eslint src/. --ext .ts -c ./.eslintrc.js",
+        "code-style-check": "npm run lint-check && npm run prettier:check",
+        "prepare": "husky install"
     },
     "author": "jkanchelov",
     "license": "MIT",
@@ -29,12 +28,14 @@
         "pixi.js": "^7.3.2"
     },
     "devDependencies": {
-        "@swc/core": "^1.3.101",
+        "@babel/core": "7.23.6",
+        "@babel/preset-env": "7.23.6",
+        "@babel/runtime": "7.23.6",
+        "babel-loader": "9.1.3",
         "@types/node": "^20.10.5",
-        "@types/jest": "^29.5.11",
         "@types/offscreencanvas": "^2019.7.3",
-        "@typescript-eslint/eslint-plugin": "^6.15.0",
-        "@typescript-eslint/parser": "^6.15.0",
+        "@typescript-eslint/eslint-plugin": "^6.16.0",
+        "@typescript-eslint/parser": "^6.16.0",
         "browserslist": "^4.22.2",
         "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.8.1",
@@ -46,17 +47,13 @@
         "html-webpack-plugin": "^5.6.0",
         "husky": "^8.0.3",
         "identity-obj-proxy": "^3.0.0",
-        "jest": "^29.7.0",
-        "jest-canvas-mock": "^2.5.2",
         "jsdom": "^23.0.1",
+        "lint-staged": "^15.2.0",
         "mini-css-extract-plugin": "^2.7.6",
         "npm-run-all": "^4.1.5",
         "prettier": "^3.1.1",
-        "pretty-quick": "^3.1.3",
         "rimraf": "^5.0.5",
-        "swc-loader": "^0.2.3",
         "terser-webpack-plugin": "^5.3.9",
-        "ts-jest": "^29.1.1",
         "ts-loader": "^9.5.1",
         "ts-node": "^10.9.2",
         "typescript": "^5.3.3",
@@ -64,22 +61,5 @@
         "webpack-cli": "^5.1.4",
         "webpack-dev-server": "^4.15.1",
         "webpack-merge": "^5.10.0"
-    },
-    "jest": {
-        "transform": {
-            ".(ts|tsx)": "ts-jest"
-        },
-        "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-        "moduleFileExtensions": [
-            "ts",
-            "tsx",
-            "js"
-        ],
-        "moduleNameMapper": {
-            ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "identity-obj-proxy"
-        },
-        "setupFiles": [
-            "jest-canvas-mock"
-        ]
     }
 }
